+++ src/passes/PostEmscripten.cpp
@@ -78,29 +78,11 @@ struct OptimizeCalls : public WalkerPass<PostWalker<OptimizeCalls>> {
 
 struct PostEmscripten : public Pass {
   void run(PassRunner* runner, Module* module) override {
-    // Optimize imports
-    optimizeImports(runner, module);
-
     // Optimize calls
     OptimizeCalls().run(runner, module);
 
     // Optimize exceptions
     optimizeExceptions(runner, module);
-  }
-
-  void optimizeImports(PassRunner* runner, Module* module) {
-    // Calling emscripten_longjmp_jmpbuf is the same as emscripten_longjmp.
-    Name EMSCRIPTEN_LONGJMP("emscripten_longjmp");
-    Name EMSCRIPTEN_LONGJMP_JMPBUF("emscripten_longjmp_jmpbuf");
-    ImportInfo info(*module);
-    auto* emscripten_longjmp =
-      info.getImportedFunction(ENV, EMSCRIPTEN_LONGJMP);
-    auto* emscripten_longjmp_jmpbuf =
-      info.getImportedFunction(ENV, EMSCRIPTEN_LONGJMP_JMPBUF);
-    if (emscripten_longjmp && emscripten_longjmp_jmpbuf) {
-      // Both exist, so it is worth renaming so that we have only one.
-      emscripten_longjmp_jmpbuf->base = EMSCRIPTEN_LONGJMP;
-    }
   }
 
   // Optimize exceptions (and setjmp) by removing unnecessary invoke* calls.
