+++ test/lld/duplicate_imports.wat.out
@@ -1,11 +1,13 @@
 (module
+ (type $i32_f32_f64_=>_f32 (func (param i32 f32 f64) (result f32)))
+ (type $i32_f64_f64_=>_f32 (func (param i32 f64 f64) (result f32)))
  (type $none_=>_none (func))
  (type $none_=>_i32 (func (result i32)))
  (type $i32_=>_i32 (func (param i32) (result i32)))
  (type $i32_i32_=>_i32 (func (param i32 i32) (result i32)))
  (type $i64_=>_i32 (func (param i64) (result i32)))
- (type $i32_f32_f64_=>_f32 (func (param i32 f32 f64) (result f32)))
- (type $i32_f64_f64_=>_f32 (func (param i32 f64 f64) (result f32)))
+ (type $f32_f64_=>_f32 (func (param f32 f64) (result f32)))
+ (type $f64_f64_=>_f32 (func (param f64 f64) (result f32)))
  (import "env" "puts" (func $puts1 (param i32) (result i32)))
  (import "env" "invoke_ffd" (func $invoke_ffd (param i32 f32 f64) (result f32)))
  (import "env" "invoke_ffd" (func $invoke_ffd2 (param i32 f64 f64) (result f32)))
@@ -21,6 +23,8 @@
  (export "main" (func $main))
  (export "__heap_base" (global $global$1))
  (export "__data_end" (global $global$2))
+ (export "dynCall_fdd" (func $dynCall_fdd))
+ (export "dynCall_ffd" (func $dynCall_ffd))
  (func $main (result i32)
   (drop
    (call $puts1
@@ -31,6 +35,20 @@
  )
  (func $__wasm_call_ctors
   (nop)
+ )
+ (func $dynCall_fdd (param $fptr i32) (param $0 f64) (param $1 f64) (result f32)
+  (call_indirect (type $f64_f64_=>_f32)
+   (local.get $0)
+   (local.get $1)
+   (local.get $fptr)
+  )
+ )
+ (func $dynCall_ffd (param $fptr i32) (param $0 f32) (param $1 f64) (result f32)
+  (call_indirect (type $f32_f64_=>_f32)
+   (local.get $0)
+   (local.get $1)
+   (local.get $fptr)
+  )
  )
  (func $legalfunc$puts2 (param $0 i64) (result i32)
   (call $legalimport$puts2
@@ -61,7 +79,9 @@
   ],
   "exports": [
     "__wasm_call_ctors",
-    "main"
+    "main",
+    "dynCall_fdd",
+    "dynCall_ffd"
   ],
   "namedGlobals": {
     "__heap_base" : "66128",
