+++ test/passes/generate-dyncalls.txt
@@ -1,12 +1,16 @@
 (module
+ (type $i32_i32_i32_=>_none (func (param i32 i32 i32)))
  (type $none_=>_i32 (func (result i32)))
  (type $i32_=>_i64 (func (param i32) (result i64)))
+ (type $i32_i32_=>_none (func (param i32 i32)))
  (type $i32_=>_i32 (func (param i32) (result i32)))
  (type $i32_i32_=>_i64 (func (param i32 i32) (result i64)))
+ (import "env" "invoke_vii" (func $invoke_vii (param i32 i32 i32)))
  (table $0 2 2 funcref)
  (elem (i32.const 0) $f1 $f2)
  (export "dynCall_i" (func $dynCall_i))
  (export "dynCall_ji" (func $dynCall_ji))
+ (export "dynCall_vii" (func $dynCall_vii))
  (func $f1 (result i32)
   (i32.const 1024)
  )
@@ -21,6 +25,13 @@
  (func $dynCall_ji (param $fptr i32) (param $0 i32) (result i64)
   (call_indirect (type $i32_=>_i64)
    (local.get $0)
+   (local.get $fptr)
+  )
+ )
+ (func $dynCall_vii (param $fptr i32) (param $0 i32) (param $1 i32)
+  (call_indirect (type $i32_i32_=>_none)
+   (local.get $0)
+   (local.get $1)
    (local.get $fptr)
   )
  )
