+++ test/passes/post-emscripten.txt
@@ -176,16 +176,3 @@
   )
  )
 )
-(module
- (type $i32_i32_=>_none (func (param i32 i32)))
- (import "env" "emscripten_longjmp" (func $a (param i32 i32)))
- (import "env" "emscripten_longjmp" (func $b (param i32 i32)))
-)
-(module
- (type $i32_i32_=>_none (func (param i32 i32)))
- (import "env" "emscripten_longjmp_jmpbuf" (func $b (param i32 i32)))
-)
-(module
- (type $i32_i32_=>_none (func (param i32 i32)))
- (import "env" "emscripten_longjmp" (func $a (param i32 i32)))
-)
